+++ src/main/groovy/org/m2ci/msp/pandocreveal/PandocRevealCompile.groovy
@@ -4,6 +4,8 @@ import org.gradle.api.DefaultTask
 import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.file.FileCollection
 import org.gradle.api.file.RegularFileProperty
+import org.gradle.api.provider.ListProperty
+import org.gradle.api.provider.MapProperty
 import org.gradle.api.tasks.*
 import org.yaml.snakeyaml.DumperOptions
 import org.yaml.snakeyaml.Yaml
@@ -35,6 +37,14 @@ class PandocRevealCompile extends DefaultTask {
     @Optional
     @InputFile
     final RegularFileProperty cslFile = project.objects.fileProperty()
+
+    @Optional
+    @Input
+    final ListProperty pandocFilters = project.objects.listProperty(String)
+
+    @Optional
+    @Input
+    final MapProperty pandocEnvironment = project.objects.mapProperty(String, String)
 
     @OutputDirectory
     final DirectoryProperty destDir = project.objects.directoryProperty()
@@ -86,8 +96,16 @@ class PandocRevealCompile extends DefaultTask {
                     '--csl', cslFile.get().asFile
             ]
         }
+        if (pandocFilters.get()) {
+            pandocFilters.get().each { filter ->
+                command += [
+                        '--filter', filter
+                ]
+            }
+        }
         project.exec {
             commandLine command
+            environment.putAll pandocEnvironment.get()
         }
     }
 }